requirements.txt
Flask==3.0.0
gunicorn==21.2.0
psycopg2-binary
authlib
requests
beautifulsoup4
google-genai
